% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annual_return.R
\name{annual_return}
\alias{annual_return}
\title{A function that calculates an annualized log-return of a given stock}
\usage{
annual_return(
  symbol,
  start_date = Sys.Date() - 365,
  end_date = Sys.Date(),
  price_type = "close",
  output_type = "percent",
  rounding = 2
)
}
\arguments{
\item{symbol}{Stock symbol.}

\item{start_date}{Start date for gathering the stock prices - default set as one year ago.}

\item{end_date}{End date for gathering the stock prices - default set as the latest available day.}

\item{price_type}{Type of prices that are to be analyzed - decide between 'open' and 'close' (default 'close').}

\item{output_type}{States how the output is expressed - decide between 'percent' (result presented in percentage) and 'float' (result not modified) - default set as 'percent'.}

\item{rounding}{The level of rounding for the output - default set as 2 decimal places.}
}
\value{
The annualized log-return of the stock is returned - it is a single numerical value.
}
\description{
A function that calculates an annualized log-return of a given stock
}
\examples{
annual_return('AAPL')
annual_return('MSFT', '2020-01-01', '2022-03-31', 'open', 'float', 4)
}
\author{
Kamil Golis
}
