% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ES_view.R
\name{ES_view}
\alias{ES_view}
\title{A function that shows Expected Shortfall defined for a given stock, time horizon, significance level and type of calculation (expressed in \%).}
\usage{
ES_view(
  symbol,
  start_date = Sys.Date() - 365,
  end_date = Sys.Date(),
  price_type = "close",
  rounding = 2,
  p = 0.05,
  h = 1,
  es_type = "hist"
)
}
\arguments{
\item{symbol}{Stock symbol.}

\item{start_date}{Start date for gathering the stock prices - default set as one year ago.}

\item{end_date}{End date for gathering the stock prices - default set as the latest available day.}

\item{price_type}{Type of prices that are to be analyzed - decide between 'open' and 'close' (default 'close').}

\item{rounding}{The level of rounding for the output - default set as 2 decimal places.}

\item{p}{Level of significance as decimal - default set as 0.05 (5\%).}

\item{h}{Time horizon for the Expected Shortfall calculation - epxressed in days - default set as 1.}

\item{es_type}{Type of Expected Shortfall calculation - decide between 'hist' (historical simulation), 'norm' (assuming Normal Distribution), 'tstud' (assuming t-Student Distribution) - default set as 'hist'.}
}
\value{
A 1-element data frame informing about the Expected Shortfall expressed in \% is returned.
}
\description{
A function that shows Expected Shortfall defined for a given stock, time horizon, significance level and type of calculation (expressed in \%).
}
\examples{
ES_view('MSFT')
ES_view('AAPL', '2019-02-01', '2021-03-31', 'open', 3, 0.05, 10, 'norm')
}
\author{
Kamil Golis
}
