% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_returns.R
\name{get_returns}
\alias{get_returns}
\title{A function that calculates daily log-returns for a given stock over a given time horizon}
\usage{
get_returns(
  symbol,
  start_date = Sys.Date() - 365,
  end_date = Sys.Date(),
  price_type = "close"
)
}
\arguments{
\item{symbol}{Stock symbol}

\item{start_date}{Start date for gathering the stock prices - default set as one year ago}

\item{end_date}{End date for gathering the stock prices - default set as the latest available day}

\item{price_type}{Type of prices that are to be analyzed - decide between 'open' and 'close' (default 'close')}
}
\value{
An xts object with daily log-returns of the stock is returned.
}
\description{
A function that calculates daily log-returns for a given stock over a given time horizon
}
\examples{
get_returns('AAPL')
get_returns('MSFT', '2020-01-01', '2022-03-31', 'open')
}
\author{
Kamil Golis
}
