% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_stock_names.R
\name{get_stock_names}
\alias{get_stock_names}
\title{A function that produces a named vector with over 1000 stocks}
\usage{
get_stock_names()
}
\value{
A named vector with stock symbols (names = stock names), that may be used later for the analysis is returned.
}
\description{
A function that produces a named vector with over 1000 stocks
}
\examples{
get_stock_names()
}
\author{
Kamil Golis
}
