% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rapm.R
\name{rapm}
\alias{rapm}
\title{A function that calculates Risk Adjusted Performance Measures for a given stock, assuming a certain benchmark and risk-free rate.}
\usage{
rapm(
  symbol,
  rf = 0.03,
  bench = "NDAQ",
  start_date = Sys.Date() - 365,
  end_date = Sys.Date(),
  price_type = "close",
  rounding = 2
)
}
\arguments{
\item{symbol}{Stock symbol.}

\item{rf}{Risk-free rate, default set as 0.03, it could be also passed as a vector of rates.}

\item{bench}{Benchmark used to calculate RAPM, default set as 'NDAQ'.}

\item{start_date}{Start date for gathering the stock prices - default set as one year ago.}

\item{end_date}{End date for gathering the stock prices - default set as the latest available day.}

\item{price_type}{Type of prices that are to be analyzed - decide between 'open' and 'close' (default 'close').}

\item{rounding}{The level of rounding for the output - default set as 2 decimal places.}
}
\value{
A data frame informing about the Risk Adjusted Performance Measures (Treynor Ratio, Tracking Error, Information Ratio) is returned.
}
\description{
A function that calculates Risk Adjusted Performance Measures for a given stock, assuming a certain benchmark and risk-free rate.
}
\examples{
rapm('MSFT')
rapm('AAPL', 0.05, 'NDAQ', '2019-02-01', '2021-03-31', 'open', 3)
}
\author{
Kamil Golis
}
