% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_price.R
\name{show_price}
\alias{show_price}
\title{A function that plots quotations for a given stock.}
\usage{
show_price(
  symbol,
  start_date = Sys.Date() - 365,
  end_date = Sys.Date(),
  price_type = "close"
)
}
\arguments{
\item{symbol}{Stock symbol.}

\item{start_date}{Start date for gathering the stock prices - default set as one year ago.}

\item{end_date}{End date for gathering the stock prices - default set as the latest available day.}

\item{price_type}{Type of prices that are to be analyzed - decide between 'open' and 'close' (default 'close').}
}
\value{
A plot depicting price levels of the stock is returned.
}
\description{
A function that plots quotations for a given stock.
}
\examples{
show_price('AAPL')
show_price('MSFT', '2020-01-01', '2022-03-31', 'open')
}
\author{
Kamil Golis
}
