% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sum_stats_stock.R
\name{sum_stats_stock}
\alias{sum_stats_stock}
\title{A function that calculates summary statistics for a given stock.}
\usage{
sum_stats_stock(
  symbol,
  start_date = Sys.Date() - 365,
  end_date = Sys.Date(),
  price_type = "close",
  rounding = 2,
  output_type
)
}
\arguments{
\item{symbol}{Stock symbol.}

\item{start_date}{Start date for gathering the stock prices - default set as one year ago.}

\item{end_date}{End date for gathering the stock prices - default set as the latest available day.}

\item{price_type}{Type of prices that are to be analyzed - decide between 'open' and 'close' (default 'close').}

\item{rounding}{The level of rounding for the output - default set as 2 decimal places.}

\item{output_type}{States how the output is expressed - decide between 'percent' (result presented in percentage) and 'float' (result not modified) - default set as 'percent'.}
}
\value{
A data frame informing about the Risk Adjusted Performance Measures (Treynor Ratio, Sharpe Ratio, Tracking Error, Information Ratio) is returned.
}
\description{
A function that calculates summary statistics for a given stock.
}
\examples{
sum_stats_stock('MSFT')
sum_stats_stock('AAPL', '2019-02-01', '2021-03-31', 'open', 3, 'float')
}
\author{
Kamil Golis
}
