% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VaR_interpret.R
\name{VaR_interpret}
\alias{VaR_interpret}
\title{A function that shows an interpretation for Value at Risk defined for a given stock, time horizon, significance level and type of calculation.}
\usage{
VaR_interpret(
  symbol,
  start_date = Sys.Date() - 365,
  end_date = Sys.Date(),
  price_type = "close",
  rounding = 2,
  p = 0.05,
  h = 1,
  var_type = "hist"
)
}
\arguments{
\item{symbol}{Stock symbol.}

\item{start_date}{Start date for gathering the stock prices - default set as one year ago.}

\item{end_date}{End date for gathering the stock prices - default set as the latest available day.}

\item{price_type}{Type of prices that are to be analyzed - decide between 'open' and 'close' (default 'close').}

\item{rounding}{The level of rounding for the output - default set as 2 decimal places.}

\item{p}{Level of significance as decimal - default set as 0.05 (5\%).}

\item{h}{Time horizon for the Expected Shortfall calculation - epxressed in days - default set as 1.}

\item{var_type}{Type of Value at Risk calculation - decide between 'hist' (historical simulation), 'norm' (assuming Normal Distribution), 'tstud' (assuming t-Student Distribution) - default set as 'hist'.}
}
\value{
The interpretation in a form of string (message) is returned.
}
\description{
A function that shows an interpretation for Value at Risk defined for a given stock, time horizon, significance level and type of calculation.
}
\examples{
VaR_interpret('MSFT')
VaR_interpret('AAPL', '2019-02-01', '2021-03-31', 'open', 3, 0.01, 10, 'norm')
}
\author{
Kamil Golis
}
