% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VaR_t.R
\name{VaR_t}
\alias{VaR_t}
\title{A function that calculates Value at Risk and Expected Shortfall assuming t-Student distribution - defined for a given stock, time horizon and significance level.}
\usage{
VaR_t(
  symbol,
  start_date = Sys.Date() - 365,
  end_date = Sys.Date(),
  price_type = "close",
  output_type = "percent",
  rounding = 2,
  p = 0.05,
  h = 1,
  df = 5
)
}
\arguments{
\item{symbol}{Stock symbol.}

\item{start_date}{Start date for gathering the stock prices - default set as one year ago.}

\item{end_date}{End date for gathering the stock prices - default set as the latest available day.}

\item{price_type}{Type of prices that are to be analyzed - decide between 'open' and 'close' (default 'close').}

\item{output_type}{States how the output is expressed - decide between 'percent' (result presented in percentage) and 'float' (result not modified) - default set as 'percent'.}

\item{rounding}{The level of rounding for the output - default set as 2 decimal places.}

\item{p}{Level of significance as decimal - default set as 0.05 (5\%).}

\item{h}{Time horizon for the Expected Shortfall calculation - epxressed in days - default set as 1.}

\item{df}{Degrees of freedom - default set as 5.}
}
\value{
A 2-element list containing Value at Risk and Expected Shortfall is returned.
}
\description{
A function that calculates Value at Risk and Expected Shortfall assuming t-Student distribution - defined for a given stock, time horizon and significance level.
}
\examples{
VaR_t('AAPL')
VaR_t('MSFT', '2019-02-01', '2021-03-31', 'open', 'float', 3, 0.03, 10, 3)
}
\author{
Kamil Golis
}
